refactor(sdk-backward-compatibility): support the backward compatibility
APIMatic Java SDK previously contained the JsonObject and JsonValue. We moved these classes to the core library, so already generated SDKs were unable to resolve the dependency of that classes. Now, core library classes will be extended in the APIMatic Java SDK. Also updated the minor version of core library.

closes #31