refactor(sdk-backward-compatibility): support the backward compatibility for JsonObject and JsonValue (#32)
APIMatic Java SDK previously contained the JsonObject and JsonValue. We moved these classes to the core library, so already generated SDKs were unable to resolve the dependency of that classes. Now, core library classes will be extended in the APIMatic Java SDK. Also updated the minor version of the core library.

closes #31

Co-authored-by: Muhammad Sufyan <muhammad.sufyan@apimatic.io>