ci(add-code-coverage-ci-script): script to upload the code coverage (#35)
Code climate tells you the key metrics of your code. Added the CI script to upload the coverage on code climate which will later be used for badges. Added the functionality in which we skip the execution of this script if the main branch's last commit contains 'ci skip' words.

closes #34