ci(checkstyle-rules-modification): configure linter for the project  (#37)
Update the rules of Java linter. Added configuration of checking the first sentence.  DesignForExtension Rule will ignore the annotation of Test and override. Add the configuration to ignore the properties of java documentation and modify the visibility of final fields. Allowed Test, override, and deprecated annotations to miss Javadoc. Allowed 16 parameters in the constructor

closes #36