ci(add-release-ci-script): add the release ci script to publish maven [skip ci] (#43)
A ci script is added for the release maven package which takes the release version as input. Create the release tag against that and update the pom file as well. Used the GitHub action to automatically create the changelog and use the input title, as the title of the changelog. Removed installation guide from Readme.   Commit and push the pom changes after successfully release.  Create Tag and changelog on the successful release

closes #42