feat(support-for-error-templates): add the support of error templates
Exception from the server is significant. From server exception, we can take critical decisions. If the exception message from the server is not self-explanatory then it is hard for a client to respond accordingly.  We noticed that we can improve the message thrown to the client from the server.  Now, the client will register an error template against a particular code.  On throwing an exception, the core library will replace the placeholder using JsonPointer.  The user will get the proper message that he wanted.

closes #52